Move `get_sdfg_path` closer to it's use in the `builder`
