+++ ndsl/dsl/dace/builder/builder.py
@@ -10,7 +10,7 @@ from dace.frontend.python.parser import DaceProgram
 import ndsl.dsl.dace.replacements  # noqa # We load in the DaCe replacements
 from ndsl import OptimizationConfig, ndsl_log
 from ndsl.comm.mpi import MPI
-from ndsl.dsl.caches.dace import get_sdfg_path
+from ndsl.dsl.dace.builder.cache import get_sdfg_path
 from ndsl.dsl.dace.builder.optimize import optimize_full_program_sdfg
 from ndsl.dsl.dace.dace_config import (
     DaceConfig,
